backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-csp>=3.7
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0
django-axes>=6.1.1
django-redis>=5.4.0
redis>=5.0.1
drf-spectacular>=0.26.5
django-fsm>=2.8.1
django-auditlog>=2.3.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
python-decouple>=3.8
gunicorn>=21.2.0
python-dateutil>=2.8.2
